+++ dpckan/functions.py
@@ -169,28 +169,60 @@ def dataset_diff(ckan_instance, datapackage, dataset):
             click.echo(f"Diferenças nos metadados recurso: {name}")
             resource_update_datastore_metadata(ckan_instance, datapackage_local_resources[name], datapackage.get_resource(name))
             update_datapackage_json_resource(ckan_instance, datapackage, datapackage_local_resources['datapackage.json'])
+            dataset_patch(ckan_instance, datapackage)
       else:
         print(f'Recurso {name} inexistente localmente, `dpckan resource delete` para excluí-lo da instância {ckan_instance.address}')
+    if name == 'datapackage.json':
+      local_data_hash = resource_hash(datapackage, name)
+      remote_data_hash = resource_url_hash(ckan_instance, dataset_remote_resources[name])
+      if local_data_hash != remote_data_hash:
+        click.echo(f"Diferenças nos metadados do conjunto: {dataset.name}")
+        update_datapackage_json_resource(ckan_instance, datapackage, datapackage_local_resources['datapackage.json'])
+        dataset_patch(ckan_instance, datapackage)
   # Compare resources in datapackage.json local with remote
   for name in datapackage.resource_names:
     if name != 'datapackage.json':
       if name not in dataset.resource_names:
         print(f'Recurso {name} não existe na instância {ckan_instance.address}. utilize `dpckan resource create` para cria-lo')
 
+def dataset_patch(ckan_instance, datapackage):
+  ckan_datapackage = frictionless_to_ckan(datapackage)
+  ckan_datapackage['id'] = datapackage.name
+  ckan_instance.call_action('package_patch', ckan_datapackage)
+
 def resource_hash(datapackage, name):
-  basepath = find_dataset_basepath(datapackage)
-  resource_path = datapackage.get_resource(name)['path']
+  resource_content = ''
   md5_hash = hashlib.md5()
-  resource_content = open(f'{basepath}/{resource_path}', "rb").read()
+  if name == 'datapackage.json':
+    ckan_datapackage = frictionless_to_ckan(datapackage)
+    resource_content = json.dumps(ckan_datapackage).encode('utf-8')
+  else:
+    basepath = find_dataset_basepath(datapackage)
+    resource_path = datapackage.get_resource(name)['path']
+    resource_content = open(f'{basepath}/{resource_path}', "rb").read()
   md5_hash.update(resource_content)
   resource_hash = md5_hash.hexdigest()
   return resource_hash
 
 def resource_url_hash(ckan_instance, resource_id):
+  resource_content = ''
   md5_hash = hashlib.md5()
   ckan_datapackage_resource = ckan_instance.action.resource_show(id = resource_id)
-  resouce_content = urlopen(ckan_datapackage_resource['url']).read()
-  md5_hash.update(resouce_content)
+  if ckan_datapackage_resource['name'] == 'datapackage.json':
+    # Buscar os metatados do dataset para retirar a key notes
+    ckan_datapackage = ckan_instance.action.package_show(id=ckan_datapackage_resource['package_id'])
+    # Buscar arquivo datapackage.json remoto para criar hash
+    resource_content = urlopen(ckan_datapackage_resource['url']).read()
+    resource_content = json.loads(resource_content.decode('utf-8'))
+    resource_content = Package(resource_content)
+    # Convert para metadados ckan para igualar à conversão do arquivo local
+    # Esta conversão é importante para comparar modificações README, CHANGELOG e CONTRIBUTING
+    resource_content = frictionless_to_ckan(resource_content)
+    resource_content['notes'] = ckan_datapackage['notes']
+    resource_content = json.dumps(resource_content).encode('utf-8')
+  else:
+    resource_content = urlopen(ckan_datapackage_resource['url']).read()
+  md5_hash.update(resource_content)
   resource_hash = md5_hash.hexdigest()
   return resource_hash
 
@@ -204,11 +236,11 @@ def frictionless_to_ckan(datapackage):
     dataset["notes"] = ""
   if os.path.isfile(README_path):
     dataset["notes"] = ""
-    dataset["notes"] += f"\n{open(README_path).read()}"
+    dataset["notes"] += f"\n{open(README_path, encoding='utf-8').read()}"
   if os.path.isfile(CONTRIBUTING_path):
-    dataset["notes"] += f"\n{open(CONTRIBUTING_path).read()}"
+    dataset["notes"] += f"\n{open(CONTRIBUTING_path, encoding='utf-8').read()}"
   if os.path.isfile(CHANGELOG_path):
-    dataset["notes"] += f"\n{open(CHANGELOG_path).read()}"
+    dataset["notes"] += f"\n{open(CHANGELOG_path, encoding='utf-8').read()}"
   if 'id' in dataset.keys():
     dataset.update({ "id" : datapackage.name})
   return dataset